+++ src/main/groovy/com/bitwiseor/log/core/repository/MemoryBacklogRepository.groovy
@@ -4,16 +4,16 @@ import com.bitwiseor.log.core.domain.BacklogEntry
 import com.bitwiseor.log.core.exception.RepositoryException
 
 class MemoryBacklogRepository implements BacklogRepository {
-	private Map<Long,BacklogEntry> entries
+	private Map<Integer,BacklogEntry> entries
 
-	MemoryBacklogRepository(final Map<Long,BacklogEntry> entries) {
+	MemoryBacklogRepository(final Map<Integer,BacklogEntry> entries) {
 		this.entries = Collections.unmodifiableMap(entries)
 	}
 	
 	@Override
-	public BacklogEntry read(Long id) {
+	public BacklogEntry read(Integer id) {
 		if(entries.containsKey(id)) {
-		return entries[id]
+			return entries[id]
 		} else {
 			throw new RepositoryException("Entry ${id} does not exist in repository")
 		}
@@ -28,7 +28,7 @@ class MemoryBacklogRepository implements BacklogRepository {
 	public void create(BacklogEntry item) {
 		item.id = nextId()
 		if(!entries.containsKey(item.id)) {
-			def modifiable = new HashMap<Long,BacklogEntry>(entries)
+			def modifiable = new HashMap<Integer,BacklogEntry>(entries)
 			modifiable[item.id] = item
 			this.entries = Collections.unmodifiableMap(modifiable)
 		} else {
@@ -39,7 +39,7 @@ class MemoryBacklogRepository implements BacklogRepository {
 	@Override
 	public void update(BacklogEntry item) {
 		if(entries.containsKey(item.id)) {
-			def modifiable = new HashMap<Long,BacklogEntry>(entries)
+			def modifiable = new HashMap<Integer,BacklogEntry>(entries)
 			modifiable[item.id] = item
 			this.entries = Collections.unmodifiableMap(modifiable)
 		} else {
@@ -48,9 +48,9 @@ class MemoryBacklogRepository implements BacklogRepository {
 	}
 
 	@Override
-	public void delete(Long id) {
+	public void delete(Integer id) {
 		if(entries.containsKey(id)) {
-			def modifiable = new HashMap<Long,BacklogEntry>(entries)
+			def modifiable = new HashMap<Integer,BacklogEntry>(entries)
 			modifiable.remove(id)
 			this.entries = Collections.unmodifiableMap(modifiable)
 		} else {
@@ -58,7 +58,7 @@ class MemoryBacklogRepository implements BacklogRepository {
 		}
 	}
 
-	private Long nextId() {
+	private Integer nextId() {
 		return entries.keySet().max() + 1
 	}
 }
